#96 Create gateway and microservice handler

+++ arkham-city-core/src/modules/project/app/app.service.ts
@@ -1,18 +1,21 @@
-import { Injectable } from '@nestjs/common';
-import { InjectModel } from '@nestjs/mongoose';
-import { App, APP_TYPE } from './app.type';
-import { Model } from 'mongoose';
-import { JWTPayload } from 'src/modules/auth/auth.type';
+import { Injectable } from "@nestjs/common";
+import { InjectModel } from "@nestjs/mongoose";
+import { App, APP_TYPE } from "./app.type";
+import { Model } from "mongoose";
+import { JWTPayload } from "src/modules/auth/auth.type";
 import {
   BadMicroserviceResponse,
   MicroserviceErrorCode,
-} from 'src/core/microservice/microservice.type';
-import { randomBytes, randomUUID } from 'crypto';
+  SuccessMicroserviceResponse,
+} from "src/core/microservice/microservice.type";
+import { randomBytes, randomUUID } from "crypto";
+import { HashService } from "src/core/hash/hash.service";
 
 @Injectable()
 export class AppService {
   constructor(
-    @InjectModel(App.name, 'metadata') private readonly appModel: Model<App>,
+    @InjectModel(App.name, "metadata") private readonly appModel: Model<App>,
+    private readonly hashService: HashService,
   ) {}
 
   async create(
@@ -21,7 +24,7 @@ export class AppService {
     name: string,
     type: APP_TYPE,
     callback: string,
-    desctiption: string = '',
+    desctiption: string = "",
   ) {
     if (
       (
@@ -29,11 +32,18 @@ export class AppService {
           project: {
             _id: projectId,
           },
+          name: name,
+          user: {
+            username: user.username,
+          },
         })
       ).length > 0
     ) {
       return new BadMicroserviceResponse(MicroserviceErrorCode.DULICATE_APP);
     }
+    const rawSecret = randomBytes(
+      parseInt(process.env.PROJECT_APP_SECRET_KEY_LENGTH as string) ?? 48,
+    ).toString("base64");
     let app = new this.appModel({
       project: {
         _id: projectId,
@@ -43,11 +53,16 @@ export class AppService {
       callback: callback,
       description: desctiption,
       clientId: randomUUID(),
-      secretKey: randomBytes(48).toString('base64'),
-      user: user,
+      secretKey: this.hashService.hash(rawSecret),
+      user: {
+        username: user.username,
+      },
     });
     app = await app.save();
-    return app.toJSON();
+    return new SuccessMicroserviceResponse({
+      ...app.toJSON(),
+      secretKey: rawSecret,
+    });
   }
 
   async update(
@@ -57,14 +72,16 @@ export class AppService {
     name: string,
     type: APP_TYPE,
     callback: string,
-    desctiption: string = '',
+    desctiption: string = "",
   ) {
     let app = await this.appModel.findOne({
       _id: appId,
       project: {
         _id: projectId,
       },
-      user: user,
+      user: {
+        username: user.username,
+      },
     });
     if (!app) {
       return new BadMicroserviceResponse(MicroserviceErrorCode.APP_NOT_FOUND);
@@ -74,7 +91,9 @@ export class AppService {
     app.callback = callback;
     app.description = desctiption;
     app = await app.save();
-    return app.toJSON();
+    return new SuccessMicroserviceResponse({
+      ...app.toJSON(),
+    });
   }
 
   async delete(user: JWTPayload, projectId: string, appId: string) {
@@ -83,11 +102,50 @@ export class AppService {
       project: {
         _id: projectId,
       },
-      user: user,
+      user: {
+        username: user.username,
+      },
     });
     if (!app) {
       return new BadMicroserviceResponse(MicroserviceErrorCode.APP_NOT_FOUND);
     }
-    await app.deleteOne();
+    app.activated = false;
+    if (
+      !process.env.CORE_HARD_DELETE ||
+      process.env.CORE_HARD_DELETE === "true"
+    ) {
+      await app.save();
+    } else {
+      await app.deleteOne();
+    }
+    return new SuccessMicroserviceResponse(true);
+  }
+
+  async all(user: JWTPayload, projectId: string) {
+    const apps = await this.appModel.find({
+      project: {
+        _id: projectId,
+      },
+      user: {
+        username: user.username,
+      },
+    });
+    return new SuccessMicroserviceResponse(apps.map((e) => e.toJSON()));
+  }
+
+  async get(user: JWTPayload, projectId: string, appId: string) {
+    const app = await this.appModel.findOne({
+      project: {
+        _id: projectId,
+      },
+      user: {
+        username: user.username,
+      },
+      _id: appId,
+    });
+    if (!app) {
+      return new BadMicroserviceResponse(MicroserviceErrorCode.APP_NOT_FOUND);
+    }
+    return new SuccessMicroserviceResponse(app.toJSON());
   }
 }
